requirements.txt
google-adk
pdfplumber
pyyaml
python-dotenv
litellm
